refactor(notifications): split panel into per-component files
Break notification-panel.tsx into a directory following the index.tsx
pattern: pure formatters + types in utils.ts, content kinds in their
own files (alert/cluster/report), plus link-text and severity-icon
helpers. Title becomes "Daily events summary" / "N-day events summary"
based on report period.

Adds left-edge slide animation via framer-motion AnimatePresence and
moves the scroll container to the shared ScrollArea. Forces Radix's
internal display: table viewport wrapper to display: block to prevent
unbreakable strings (URLs, IDs) from being clipped past the right edge.
